# sstable: use bytealloc.A in reader, writer, suffix rewriter